Extract ValueTransformer into its own file
